260310_promotion_planning/executor/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pandas==2.3.3
numpy>=1.24.0
python-multipart==0.0.20
azure-storage-blob>=12.27.0
azure-identity>=1.13.0
